requirements.txt
reportlab>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
Pillow>=10.0.0
pywin32>=306; platform_system=="Windows"
